Move getWorkspaceInfo into utils folder
